Fix text rendering: when a lot of text is rendered, the result becomes a mess. Caused by the overflow of IndexBuffer, which is a list of short. And when the index goes beyond 65535, a invalid index is send to GPU. Solution: change short to int. Rewrite button in the ImGui way (GUIState.cs). State updating is still buggy.

+++ IMGUI/DrawList/DrawList.Bezier.cs
@@ -9,7 +9,7 @@ namespace ImGui
     {
         private int _bezier_vtxWritePosition;
         private int _bezier_idxWritePosition;
-        private short _bezier_currentIdx;
+        private int _bezier_currentIdx;
 
         private List<int> _BezierControlPointIndex = new List<int>();
 
@@ -19,9 +19,9 @@ namespace ImGui
             _bezier_vtxWritePosition++;
         }
 
-        private void AppendBezierIndex(short offsetToCurrentIndex)
+        private void AppendBezierIndex(int offsetToCurrentIndex)
         {
-            bezierIndexBuffer[_bezier_idxWritePosition] = new DrawIndex { Index = (short)(_bezier_currentIdx + offsetToCurrentIndex) };
+            bezierIndexBuffer[_bezier_idxWritePosition] = new DrawIndex { Index = _bezier_currentIdx + offsetToCurrentIndex };
             _bezier_idxWritePosition++;
         }
